Modified the javascript to reload the site with url parameters that define where we've drilled until now. This creates a unique url for each drilldown.
Can be improved (e.g. the alert when no children are present or let user know where he/she is at without having to 'parse' it from the url parameters.

This introduces a dependency on purl.js (jQuery based url parser). It could be implemented easily but why go through the extra work?