Create sandbox for researching potential options
I've been working on some new d3 charts away from the structure of this
app. I'm moving that directory into a `sandbox` folder so the work can
be better followed and to allow others to toy with the charts as well.

Currently I've duplicated a lot of our data so I can make breaking
changes and not worry about our 'production' versions. I've also started
on a simplified set of data because I'm so unfamiliar with our state
data that it's hard for me to notice issues.