Update Readme
I'm moving the sandbox information to the top of the README since that's where most of the current work is happening. I've also added a Contributing section to try and make it a bit easier to understand.

Removed the Publishing section since we aren't currently at a place we're running the site anywhere and those instructions were really kinda complicated. Once we get to a place we're ready to actually publish the site somewhere we can readdress those instructions.