Display "compared" budgets on right/bottom
This was a minor UI nag that I noticed during the OpenBudget demo: I
found myself wanting to compare the budget in a left-to-right fashion
(or top-to-bottom) chronologically.

Since the default option for the comparison tool is comparing FY17-18 to
FY16-17, I'd argue that we should display the bars such that FY16-17 is
to the left or top, and FY17-18 is to the right or bottom by default.

(The color coding helps quite a bit in understanding what happens when
the budget to compare against is changed.)

The one gotcha here is that the colors array is given by D3. I was going
to just switch the default budgets around, but I like that the current
budget is in green rather than orange.